+++ capreolus/trainer/test_trainer.py
@@ -3,7 +3,7 @@ import numpy as np
 import os
 import tensorflow as tf
 from capreolus.benchmark import DummyBenchmark
-from capreolus.extractor import EmbedText
+from capreolus.extractor.embedtext import EmbedText
 from capreolus.sampler import TrainDataset
 from capreolus.trainer import TensorFlowTrainer
